+++ src/botlib/l_script.h
@@ -186,7 +186,7 @@ typedef struct token_s
 	int type;                       //last read token type
 	int subtype;                    //last read token sub type
 #ifdef NUMBERVALUE
-	unsigned long int intvalue; //integer value
+	uint32_t intvalue; //integer value
 	long double floatvalue;         //floating point value
 #endif //NUMBERVALUE
 	char *whitespace_p;             //start of white space before token
@@ -242,7 +242,7 @@ void StripDoubleQuotes( char *string );
 //remove any leading and trailing single quotes from the token
 void StripSingleQuotes( char *string );
 //read a possible signed integer
-signed long int ReadSignedInt( script_t *script );
+int32_t ReadSignedInt( script_t *script );
 //read a possible signed floating point number
 long double ReadSignedFloat( script_t *script );
 //set an array with punctuations, NULL restores default C/C++ set
