+++ src/game/be_ai_chat.h
@@ -102,13 +102,13 @@ void BotGetChatMessage( int chatstate, char *buf, int size );
 //checks if the first string contains the second one, returns index into first string or -1 if not found
 int StringContains( char *str1, char *str2, int casesensitive );
 //finds a match for the given string
-int BotFindMatch( char *str, bot_match_t *match, unsigned long int context );
+int BotFindMatch( char *str, bot_match_t *match, uint32_t context );
 //returns a variable from a match
 void BotMatchVariable( bot_match_t *match, int variable, char *buf, int size );
 //unify all the white spaces in the string
 void UnifyWhiteSpaces( char *string );
 //replace all the context related synonyms in the string
-void BotReplaceSynonyms( char *string, unsigned long int context );
+void BotReplaceSynonyms( char *string, uint32_t context );
 //loads a chat file for the chat state
 int BotLoadChatFile( int chatstate, char *chatfile, char *chatname );
 //store the gender of the bot in the chat state
