+++ src/game/q_math.cpp
@@ -598,13 +598,13 @@ void VectorRotate( vec3_t in, vec3_t matrix[3], vec3_t out ) {
 ** float q_rsqrt( float number )
 */
 float Q_rsqrt( float number ) {
-	long i;
+	int32_t i;
 	float x2, y;
 	const float threehalfs = 1.5F;
 
 	x2 = number * 0.5F;
 	y  = number;
-	i  = *( long * ) &y;                        // evil floating point bit level hacking
+	i  = *( int32_t * ) &y;                        // evil floating point bit level hacking
 	i  = 0x5f3759df - ( i >> 1 );               // what the fuck?
 	y  = *( float * ) &i;
 	y  = y * ( threehalfs - ( x2 * y * y ) );   // 1st iteration
